add pcm bar

+++ gprogressbar.cpp
@@ -11,6 +11,6 @@ void GProgressBar::mousePressEvent(QMouseEvent *e)
 {
     QProgressBar::mousePressEvent(e);
     double p = (double)(e->pos().x()) / (double)width();
-    int val = p*(maximum()-minimum())+minimum();
+    qint64 val = p*(maximum()-minimum())+minimum();
     emit valFromMouse(val);
 }
